beckend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
openai>=1.35.0
email-validator>=2.1.1
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.9
pytest
